+++ backend-rust/CHANGELOG.md
@@ -4,11 +4,17 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.1.30] - 2025-02-27
+
+### Fixed
+
+- Relaxing validation constraint on all database rows affecting bakers requiring a single row update prior to protocol version 7.
+
 ## [0.1.29] - 2025-02-26
 
 ### Fixed
 
-- Relaxing validation constraint on database rows affecting bakers prior to protocol version 7.
+- Relaxing validation constraint on database rows related to prepared bakers events when protocol 7.
 
 ## [0.1.28] - 2025-02-26